Relocate search components
